Separate scorecard components into individual files.
